bfdd: Move sbfd_reflector_show_info_cmd into bfdd_vty.c
This is a show command that does not belong in the bfdd_cli.c
file, as that bfdd_cli.c is for northbound code.
